backend/pyproject.toml
[project]
name = "defensefood"
version = "0.1.0"
description = "EU Food Fraud Vulnerability Intelligence System"
requires-python = ">=3.10"
dependencies = [
    "defensefood-core",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic>=2.0",
    "pandas>=2.0",
    "numpy>=1.26",
    "openpyxl>=3.1",
    "requests>=2.31",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]
